HW-3: major name refactoring
