+++ src/components/MainFooter.vue
@@ -1,7 +1,7 @@
 <script setup>
 import { ref } from 'vue';
 
-const taskNumber = ref(3);
+const taskNumber = ref(4);
 </script>
 
 <template>
