+++ src/components/Search.vue
